haku ja tiedot yhdistetty
