+++ frontend/src/components/SearchBar.css
@@ -1,4 +1,4 @@
-.search-form {
+/*.search-form {
   margin: 20px 0;
   text-align: center;
 }
@@ -26,3 +26,4 @@
 .search-form button:hover {
   background-color: #0056b3;
 }
+*/
